PineconeEmbed/requirements.txt
flask
openai
pinecone-client